+++ modules/admin/resources/js/zaa.js
@@ -1,5 +1,17 @@
 var zaa = angular.module("zaa", ["ui.router", "ngResource", "ngDragDrop", "angular-loading-bar", "ngFileUpload", "ngWig", "slugifier"]);
 
+
+/**
+ * guid creator
+ * @returns {String}
+ */
+function guid() {
+	function s4() {
+		return Math.floor((1 + Math.random()) * 0x10000).toString(16).substring(1);
+	}
+  return s4() + s4() + '-' + s4() + '-' + s4() + '-' + s4() + '-' + s4() + s4() + s4();
+}
+
 /* zephir angular admin */
 /* resolve controller: https://github.com/angular-ui/ui-router/wiki#resolve */
 (function() {